% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline.R
\name{sparkline}
\alias{sparkline}
\title{Interactive Sparklines}
\usage{
sparkline(values, ..., width = 60, height = 20, elementId = NULL,
  renderSelector = NULL)
}
\arguments{
\item{values}{\code{vector} of values to visualize}

\item{...}{additional options passed to \code{jquery.sparkline};  see
\href{http://omnipotent.net/jquery.sparkline/#common}{jquery.sparkline docs}}

\item{height, width}{height and width of sparkline htmlwidget
specified in any valid \code{CSS} size unit.}

\item{elementId}{\code{string} id as a valid \code{CSS} element id.}

\item{renderSelector}{\code{string} as \code{CSS} selector to render
in a DOM element other than the \code{htmlwidget} container}
}
\description{
Create interactive sparklines for inline visualization.
}

